Add files via upload
This script looks in the current directory for folders containing the string "wk" or "Wk". Then creates a list of the files in those folders. It also stores the number that appears after the string as a variable (week) - e.g. wk1 would store the variable "1". This will later be used to match the week to the correct section to upload files on Moodle.